Move NFT functions into libxrpl
