+++ src/ripple/app/ledger/LedgerToJson.h
@@ -25,6 +25,7 @@
 #include <ripple/basics/StringUtilities.h>
 #include <ripple/json/Object.h>
 #include <ripple/protocol/STTx.h>
+#include <ripple/protocol/serialize.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/rpc/Context.h>
 
@@ -70,22 +71,6 @@ addJson(Json::Value&, LedgerFill const&);
 Json::Value
 getJson(LedgerFill const&);
 
-/** Serialize an object to a blob. */
-template <class Object>
-Blob
-serializeBlob(Object const& o)
-{
-    Serializer s;
-    o.add(s);
-    return s.peekData();
-}
-
-/** Serialize an object to a hex string. */
-inline std::string
-serializeHex(STObject const& o)
-{
-    return strHex(serializeBlob(o));
-}
 }  // namespace ripple
 
 #endif
